Cleanup: Use wait_for_print method to ensure jobs are correctly transferred.
Log exceptions when they occur instead of during graceful exit.

Rename old service and status files to legacy_*; address review
feedback.
