+++ securedrop_export/archive.py
@@ -9,7 +9,7 @@ import tempfile
 from securedrop_export.exceptions import ExportException
 from securedrop_export.status import BaseStatus
 from securedrop_export.command import Command
-from securedrop_export.directory_util import safe_extractall
+from securedrop_export.directory import safe_extractall
 
 logger = logging.getLogger(__name__)
