+++ securedrop_export/disk/__init__.py
@@ -1,0 +1,1 @@
+from .service import Service  # noqa: F401
