+++ securedrop_export/disk/status.py
@@ -3,24 +3,27 @@ from securedrop_export.status import BaseStatus
 
 class Status(BaseStatus):
 
-    LEGACY_ERROR_GENERIC = "ERROR_GENERIC"
+    NO_DEVICE_DETECTED = "NO_DEVICE_DETECTED"
+    INVALID_DEVICE_DETECTED = (
+        "INVALID_DEVICE_DETECTED"  # Multi partitioned, not encrypted, etc
+    )
+    MULTI_DEVICE_DETECTED = "MULTI_DEVICE_DETECTED"  # Not currently supported
 
-    # Legacy USB preflight related
-    LEGACY_USB_CONNECTED = "USB_CONNECTED"  # Success
-    LEGACY_USB_NOT_CONNECTED = "USB_NOT_CONNECTED"
-    LEGACY_ERROR_USB_CHECK = "ERROR_USB_CHECK"
+    DEVICE_LOCKED = "DEVICE_LOCKED"  # One device detected, and it's locked
+    DEVICE_WRITABLE = (
+        "DEVICE_WRITABLE"  # One device detected, and it's unlocked (and mounted)
+    )
 
-    # Legacy USB Disk preflight related errors
-    LEGACY_USB_ENCRYPTED = "USB_ENCRYPTED"  # Success
-    LEGACY_USB_ENCRYPTION_NOT_SUPPORTED = "USB_ENCRYPTION_NOT_SUPPORTED"
+    ERROR_UNLOCK_LUKS = "ERROR_UNLOCK_LUKS"
+    ERROR_UNLOCK_GENERIC = "ERROR_UNLOCK_GENERIC"
+    ERROR_MOUNT = "ERROR_MOUNT"  # Unlocked but not mounted
 
-    # Can be raised during disk format check
-    LEGACY_USB_DISK_ERROR = "USB_DISK_ERROR"
-
-    # Legacy Disk export errors
-    LEGACY_USB_BAD_PASSPHRASE = "USB_BAD_PASSPHRASE"
-    LEGACY_ERROR_USB_MOUNT = "ERROR_USB_MOUNT"
-    LEGACY_ERROR_USB_WRITE = "ERROR_USB_WRITE"
-
-    # New
     SUCCESS_EXPORT = "SUCCESS_EXPORT"
+    ERROR_EXPORT = "ERROR_EXPORT"  # Could not write to disk
+
+    # export succeeds but drives were not properly unmounted
+    ERROR_EXPORT_CLEANUP = "ERROR_EXPORT_CLEANUP"
+
+    DEVICE_ERROR = (
+        "DEVICE_ERROR"  # Something went wrong while trying to check the device
+    )
