+++ tests/test_directory.py
@@ -4,10 +4,10 @@ import tempfile
 import shutil
 
 from pathlib import Path
-from securedrop_export import directory_util
+from securedrop_export import directory
 
 
-class TestDirectoryUtil:
+class TestDirectory:
 
     _REL_TRAVERSAL = "../../../whee"
     _SAFE_RELPATH = "./hi"
@@ -32,53 +32,51 @@ class TestDirectoryUtil:
 
     def test_safe_mkdir_error_base_relpath(self):
         with pytest.raises(ValueError):
-            directory_util.safe_mkdir(base_path=Path("."))
+            directory.safe_mkdir(base_path=Path("."))
 
     def test_safe_mkdir_error_basepath_path_traversal(self):
         with pytest.raises(ValueError):
-            directory_util.safe_mkdir(f"{self.homedir}{self._REL_TRAVERSAL}")
+            directory.safe_mkdir(f"{self.homedir}{self._REL_TRAVERSAL}")
 
     def test_safe_mkdir_error_relpath_path_traversal(self):
         with pytest.raises(ValueError):
-            directory_util.safe_mkdir(f"{self.homedir}", f"{self._REL_TRAVERSAL}")
+            directory.safe_mkdir(f"{self.homedir}", f"{self._REL_TRAVERSAL}")
 
     def test_safe_mkdir_success(self):
-        directory_util.safe_mkdir(f"{self.homedir}")
+        directory.safe_mkdir(f"{self.homedir}")
 
     def test_safe_mkdir_success_with_relpath(self):
-        directory_util.safe_mkdir(f"{self.homedir}", f"{self._SAFE_RELPATH}")
+        directory.safe_mkdir(f"{self.homedir}", f"{self._SAFE_RELPATH}")
 
         assert os.path.exists(f"{self.homedir}{self._SAFE_RELPATH}")
 
     def test_safe_mkdir_success_another_relpath(self):
-        directory_util.safe_mkdir(f"{self.homedir}", f"{self._SAFE_RELPATH2}")
+        directory.safe_mkdir(f"{self.homedir}", f"{self._SAFE_RELPATH2}")
 
         assert os.path.exists(f"{self.homedir}{self._SAFE_RELPATH2}")
 
     def test_safe_mkdir_weird_path(self):
         with pytest.raises(ValueError):
-            directory_util.safe_mkdir(f"{self.homedir}", f"{self._UNSAFE_RELPATH}")
+            directory.safe_mkdir(f"{self.homedir}", f"{self._UNSAFE_RELPATH}")
 
     def test__check_all_permissions_path_missing(self):
         with pytest.raises(ValueError):
-            directory_util._check_all_permissions(
-                f"{self.homedir}", f"{self._SAFE_RELPATH}"
-            )
+            directory._check_all_permissions(f"{self.homedir}", f"{self._SAFE_RELPATH}")
 
     def test_check_dir_perms_unsafe(self):
         path = Path(f"{self.homedir}{self._SAFE_RELPATH}")
 
-        directory_util.safe_mkdir(path)
+        directory.safe_mkdir(path)
 
         # Not what we want, ever
         path.chmod(0o666)
 
         with pytest.raises(RuntimeError):
-            directory_util._check_dir_permissions(path)
+            directory._check_dir_permissions(path)
 
     def test_check_all_perms_invalid_full_path(self):
         path = Path(f"{self.homedir}/idontexist")
         base = Path(f"{self.homedir}")
 
         # Returns without error
-        assert directory_util._check_all_permissions(path, base) is None
+        assert directory._check_all_permissions(path, base) is None
